requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alpaca-trade-api==3.2.0
appdirs==1.4.4
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
cloudpickle==2.2.1
cryptography==44.0.1
curl_cffi==0.11.4
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
deprecation==2.1.0
entrypoints==0.4
exchange_calendars==3.6.3
Farama-Notifications==0.0.4
FinRL==0.3.7
fonttools==4.38.0
frozendict==2.4.6
frozenlist==1.3.3
greenlet==3.2.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
html5lib==1.1
idna==3.10
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
jax-jumpy==1.0.0
jedi==0.19.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lxml==5.3.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.0
pandas==1.5.3
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
prompt_toolkit==3.0.48
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.17.2
pyluach==2.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.2.1
requests==2.31.0
six==1.17.0
soupsieve==2.4.1
SQLAlchemy==2.0.41
stable-baselines3==2.0.0
stockstats==0.6.5
toolz==0.12.1
torch==1.13.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==15.0.1
wrds==3.3.0
yarl==1.9.4
yfinance==0.2.64
zipp==3.15.0
